+++ src/main/java/com/salessparrow/api/config/FilterConfig.java
@@ -5,9 +5,23 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import com.salessparrow.api.filter.SameSiteCookieFilter;
+import com.salessparrow.api.filter.SanitizationFilter;
 
 @Configuration
 public class FilterConfig {
+
+	/**
+	 * Register SanitizationFilter
+	 * @return FilterRegistrationBean<SanitizationFilter>
+	 */
+	@Bean
+	public FilterRegistrationBean<SanitizationFilter> sanitizationFilter() {
+		FilterRegistrationBean<SanitizationFilter> registrationBean = new FilterRegistrationBean<>();
+		registrationBean.setFilter(new SanitizationFilter());
+		registrationBean.addUrlPatterns("/*");
+		registrationBean.setOrder(1);
+		return registrationBean;
+	}
 
 	/**
 	 * Register SameSiteCookieFilter
@@ -18,6 +32,7 @@ public class FilterConfig {
 		FilterRegistrationBean<SameSiteCookieFilter> registrationBean = new FilterRegistrationBean<>();
 		registrationBean.setFilter(new SameSiteCookieFilter());
 		registrationBean.addUrlPatterns("/*"); // or specific URL patterns
+		registrationBean.setOrder(2);
 		return registrationBean;
 	}