+++ src/main/java/com/salessparrow/api/lib/Util.java
@@ -64,7 +64,7 @@ public class Util {
 	}
 
 	/**
-	 * <<<<<<< HEAD Encode plain text to base64
+	 * Encode plain text to base64
 	 * @param plainText - String to be encoded
 	 * @return String - Encoded string
 	 */
@@ -158,6 +158,27 @@ public class Util {
 			return dateFormat.format(date);
 		}
 		return null;
+	}
+
+	/**
+	 * Unescape special characters in a string intended for plaintext contexts only.
+	 *
+	 * This function is intended for use only with strings that will be rendered as
+	 * plaintext. It is NOT suitable for HTML strings, XML strings, or any other context
+	 * where special characters may have syntactic meaning.
+	 *
+	 * Current Implementation: - The ampersand ("&amp;") is unescaped to "&"
+	 *
+	 * Future versions may include additional un-escaping rules as needed.
+	 * @param input The original string containing escaped special characters.
+	 * @return A new string where certain special characters have been unescaped.
+	 */
+	public String unEscapeSpecialCharactersForPlainText(String input) {
+		return input.replace("&amp;", "&");
+	}
+
+	public String replaceNewLineWithBreak(String input) {
+		return input.replace("\n", "<br>");
 	}
 
 }