+++ repo-docs/CHANGELOG.md
@@ -16,11 +16,11 @@
 - API - Delete Task in an Account Endpoint [#59](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/issues/59)
 - Add CHANGELOG.md [#62](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/issues/62)
 - Add RELEASE_PROCESS.md [#63](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/issues/63)
+- Add Sanitization Filter for Request Params, Request Body, and Headers [#87](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/issues/87)
 
 ### Bug Fixes:
 
 - Bug: DynamoDB Upsert Behavior Updates All Fields Instead of Specific Ones [#41](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/issues/41)
-- Fix: Updated status enum values [#27](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/pull/27)
 
 ### Miscellaneous:
 
@@ -46,6 +46,7 @@
 - API - Search Accounts Endpoint [#15](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/issues/15)
 - API - Create Note Endpoint [#16](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/issues/16)
 - API - Get Note By Id Endpoint [#17](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/issues/17)
+- Updated status enum values [#27](https://github.com/TrueSparrowSystems/AI-SalesSparrow-API/pull/27)
 
 ---